+++ packages/renterd/lib/src/abstract/contract_absts/acquire_a_contract_abst.dart
@@ -1,0 +1,9 @@
+import 'package:http/http.dart' as http;
+
+abstract class AcquireAContractAbst {
+  Future<http.Response> acquireAContract({
+    required String id,
+    required String duration,
+    required int priority,
+  });
+}