+++ packages/renterd/lib/src/abstract/contract_absts/add_contract_abst.dart
@@ -1,0 +1,10 @@
+import 'package:http/http.dart' as http;
+
+abstract class AddContractAbst {
+  Future<http.Response> addContract({
+    required String id,
+    required Map<dynamic, dynamic> contract,
+    required String totalCost,
+    required int startHeight,
+  });
+}