+++ packages/renterd/lib/src/abstract/contract_absts/delete_contract_by_id_abst.dart
@@ -1,0 +1,7 @@
+import 'package:http/http.dart' as http;
+
+abstract class DeleteContractByIdAbst {
+  Future<http.Response> deleteContractById({
+    required String id,
+  });
+}