+++ packages/renterd/lib/src/abstract/contract_absts/release_previous_contract_abst.dart
@@ -1,0 +1,8 @@
+import 'package:http/http.dart' as http;
+
+abstract class ReleasePreviousContractAbst {
+  Future<http.Response> releasePreviousContract({
+    required String id,
+    required int lockId,
+  });
+}