+++ packages/renterd/lib/src/apis/contracts_api.dart
@@ -1,0 +1,6 @@
+const String getContractsApi = "api/bus/contracts";
+String postContractsApi(String id) => "api/bus/contract/$id";
+String acquireAContractApi(String id) => "api/bus/contract/$id/acquire";
+String releasePreviousContractApi(String id) => "api/bus/contract/$id/release";
+String getContractByIdApi(String id) => "api/bus/contract/$id";
+String deleteContractByIdApi(String id) => "api/bus/contract/$id";