+++ packages/renterd/lib/src/logic/controllers/contract_controllers/get_contracts_controller.dart
@@ -1,0 +1,15 @@
+import 'package:http/http.dart' as http;
+
+import '../../../abstract/contract_absts/get_contracts_abst.dart';
+
+class GetContractsController {
+  final GetContractsAbst getContractsAbst;
+
+  GetContractsController({
+    required this.getContractsAbst,
+  });
+
+  Future<http.Response> call() async {
+    return await getContractsAbst.getContracts();
+  }
+}